danzeenldc_api/requirements.txt
# django rest framework
djangorestframework

# MySQL
mysqlclient

# clean html input
bleach

# filter extern requests
django-cors-headers

# JSONField for Users
jsonfield

# JWT
djangorestframework-jwt
